# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides api docs and storage management.",
        "title": "Api-Docs-Provider API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.5.5"
    },
    "basePath": "/",
    "paths": {
        "/asyncapi/docs": {
            "get": {
                "description": "Get all asyncapi docs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AsyncAPI"
                ],
                "summary": "Get docs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of asyncapi docs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/asyncapi/storage": {
            "get": {
                "description": "Get meta information of all stored items.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "AsyncAPI"
                ],
                "summary": "List storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "stored items",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AsyncapiItem"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/asyncapi/storage/{id}": {
            "put": {
                "description": "Store an asyncapi doc.",
                "consumes": [
                    "application/octet-stream"
                ],
                "tags": [
                    "AsyncAPI"
                ],
                "summary": "Store doc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "doc id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "doc",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove an asyncapi doc.",
                "tags": [
                    "AsyncAPI"
                ],
                "summary": "Delete doc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "doc id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "description": "Get basic service and runtime information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get service info",
                "responses": {
                    "200": {
                        "description": "info",
                        "schema": {
                            "$ref": "#/definitions/lib.SrvInfo"
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swagger": {
            "get": {
                "description": "Get all swagger docs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "Get docs (deprecated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "user roles",
                        "name": "X-User-Roles",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of swagger docs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swagger/docs": {
            "get": {
                "description": "Get all swagger docs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "Get docs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "user roles",
                        "name": "X-User-Roles",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of swagger docs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swagger/storage": {
            "get": {
                "description": "Get meta information of all stored items.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Swagger"
                ],
                "summary": "List storage",
                "responses": {
                    "200": {
                        "description": "stored items",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SwaggerItem"
                            }
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/swagger/storage-refresh": {
            "patch": {
                "description": "Trigger swagger docs refresh.",
                "tags": [
                    "Swagger"
                ],
                "summary": "Refresh storage",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "lib.MemStats": {
            "type": "object",
            "properties": {
                "alloc": {
                    "type": "integer"
                },
                "alloc_total": {
                    "type": "integer"
                },
                "gc_cycles": {
                    "type": "integer"
                },
                "sys_total": {
                    "type": "integer"
                }
            }
        },
        "lib.SrvInfo": {
            "type": "object",
            "properties": {
                "mem_stats": {
                    "$ref": "#/definitions/lib.MemStats"
                },
                "name": {
                    "type": "string"
                },
                "up_time": {
                    "$ref": "#/definitions/time.Duration"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.AsyncapiItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.SwaggerItem": {
            "type": "object",
            "properties": {
                "ext_paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                1,
                1000,
                1000000,
                1000000000
            ],
            "x-enum-varnames": [
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second"
            ]
        }
    }
}